projects/02-etl-pipeline-framework/requirements.txt
# Core Dependencies
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
boto3==1.34.0

# AWS SDK
awscli==1.32.0

# Data Processing
pyarrow==14.0.1
fastparquet==2023.10.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

